#76 rename: AuhtorizationRequestManager manager 디렉토리로 이동
